toy-robot-scientist/requirements.txt
console-menu==0.6.0
isodate==0.6.0
kanren==0.2.3
multipledispatch==0.6.0
Pillow==8.3.1
pyparsing==2.4.7
rdflib==6.0.0
six==1.16.0
toolz==0.11.1
unification==0.2.2
tabulate==0.8.9
